% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{HoltWintersBenchmark}
\alias{HoltWintersBenchmark}
\title{Auxiliary function for the exponential smoothing forecasting method.
From https://github.com/M4Competition/M4-methods/blob/master/Benchmarks\%20and\%20Evaluation.R}
\usage{
HoltWintersBenchmark(input, h, level = c(80, 95))
}
\arguments{
\item{input}{\itemize{
\item Input time series
}}

\item{h}{\itemize{
\item Forecasting horizon
}}

\item{level}{\itemize{
\item Numeric vector (length 2) containing the upper and lower bound for interval forecasting
}}
}
\value{
h-step forecast for the Holt-Winters classic method
}
\description{
Auxiliary function for the exponential smoothing forecasting method.
From https://github.com/M4Competition/M4-methods/blob/master/Benchmarks\%20and\%20Evaluation.R
}
