% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputVerification.R
\name{input_check_level}
\alias{input_check_level}
\title{input_check_level}
\usage{
input_check_level(level)
}
\arguments{
\item{level}{\itemize{
\item Numeric vector (length 2) containing the upper and lower bound for interval forecasting
}}
}
\value{
\itemize{
\item NULL
}
}
\description{
Verifies that the level input to the univariate forecasting function has the correct types and shape
}
