% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputVerification.R
\name{input_check_multivariate}
\alias{input_check_multivariate}
\title{input_check_multivariate}
\usage{
input_check_multivariate(X, h)
}
\arguments{
\item{X}{\itemize{
\item Target matrix (hxk) for k variables
}}

\item{h}{\itemize{
\item Forecasting horizon (numeric scalar)
}}
}
\value{
\itemize{
\item NULL
}
}
\description{
Verifies that the input values to the multivariate forecasting function have the correct types
}
