% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputVerification.R
\name{input_check_univariate}
\alias{input_check_univariate}
\title{input_check_univariate}
\usage{
input_check_univariate(x, h)
}
\arguments{
\item{x}{\itemize{
\item Input univariate series (numeric vector)
}}

\item{h}{\itemize{
\item Forecasting horizon (numeric scalar)
}}
}
\value{
\itemize{
\item NULL
}
}
\description{
Verifies that the input values to the univariate forecasting function have the correct types
}
