% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{M4_benchmarks}
\alias{M4_benchmarks}
\title{Fucntion producing the forecasts according to the benchmark methods of the M4 competition.
Adapted from https://github.com/M4Competition/M4-methods/blob/master/Benchmarks\%20and\%20Evaluation.R}
\usage{
M4_benchmarks(input, fh)
}
\arguments{
\item{input}{\itemize{
\item Input time series
}}

\item{fh}{\itemize{
\item Forecasting horizon
}}
}
\value{
List containing the forecast and the computational times of the different methods (Naive, Naive Seasonal, Naive 2, Simple Exponential smoothing ,Holt Winters, Holt Winters Damped, Theta, Comb)
}
\description{
Fucntion producing the forecasts according to the benchmark methods of the M4 competition.
Adapted from https://github.com/M4Competition/M4-methods/blob/master/Benchmarks\%20and\%20Evaluation.R
}
