% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{multiplicativeSeasonalityDecomposition}
\alias{multiplicativeSeasonalityDecomposition}
\title{multiplicativeSeasonalityDecomposition}
\usage{
multiplicativeSeasonalityDecomposition(input, fh)
}
\arguments{
\item{input}{\itemize{
\item Input time series
}}

\item{fh}{\itemize{
\item Forecasting horizon
}}
}
\value{
\itemize{
\item{\code{des_input}: }{Deseasonalized input time series}
\item{\code{SIOut}: }{Additional deseasonalization parameters}
}
}
\description{
Auxiliary function for computing the multiplicative seasonality decomposition
From \url{https://github.com/M4Competition/M4-methods/blob/master/Benchmarks\%20and\%20Evaluation.R}
}
