% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{naive2Benchmark}
\alias{naive2Benchmark}
\title{Auxiliary function for the naive 2 forecasting method.
From https://github.com/M4Competition/M4-methods/blob/master/Benchmarks\%20and\%20Evaluation.R}
\usage{
naive2Benchmark(input, h, level = c(80, 95))
}
\arguments{
\item{input}{\itemize{
\item Input time series
}}

\item{h}{\itemize{
\item Forecasting horizon
}}

\item{level}{\itemize{
\item Numeric vector (length 2) containing the upper and lower bound for interval forecasting
}}
}
\value{
h-step forecast for the theta classic method
}
\description{
Auxiliary function for the naive 2 forecasting method.
From https://github.com/M4Competition/M4-methods/blob/master/Benchmarks\%20and\%20Evaluation.R
}
