% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{naive_seasonal}
\alias{naive_seasonal}
\title{#' Auxiliary function for the Theta classic forecasting method.
From https://github.com/M4Competition/M4-methods/blob/master/Benchmarks\%20and\%20Evaluation.R}
\usage{
naive_seasonal(input, fh, level = c(80, 95))
}
\arguments{
\item{input}{\itemize{
\item Input time series (numeric vector)
}}

\item{fh}{\itemize{
\item Forecasting horizon (numeric)
}}

\item{level}{\itemize{
\item Numeric vector (length 2) containing the upper and lower bound for interval forecasting
}}
}
\value{
h-step forecast for the seasonal naive method
}
\description{
#' Auxiliary function for the Theta classic forecasting method.
From https://github.com/M4Competition/M4-methods/blob/master/Benchmarks\%20and\%20Evaluation.R
}
