% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{naiveBenchmark}
\alias{naiveBenchmark}
\title{naiveBenchmark}
\usage{
naiveBenchmark(input, h, level = c(80, 95))
}
\arguments{
\item{input}{\itemize{
\item Input time series (numeric vector)
}}

\item{h}{\itemize{
\item Forecasting horizon (numeric scalar)
}}

\item{level}{\itemize{
\item Numeric vector (length 2) containing the upper and lower bound for interval forecasting
}}
}
\value{
h-step forecast for the naive forecasting method (numeric vector - length h)
}
\description{
Auxiliary function for the naive forecasting method.
From \url{https://github.com/M4Competition/M4-methods/blob/master/Benchmarks\%20and\%20Evaluation.R}
}
\examples{
x <- AirPassengers
splitting_point <- round(2*length(x)/3)
x_train <- x[1:splitting_point]
h <- 5
x_hat <- naiveBenchmark(x_train,h)
}
