% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{thetaBenchmark}
\alias{thetaBenchmark}
\title{thetaBenchmark}
\usage{
thetaBenchmark(input, h)
}
\arguments{
\item{input}{\itemize{
\item Input time series (numeric vector)
}}

\item{h}{\itemize{
\item Forecasting horizon (numeric scalar)
}}
}
\value{
h-step forecast for the Theta method (numeric vector - length h)
}
\description{
Auxiliary function for the Theta forecasting method.
From \url{https://github.com/M4Competition/M4-methods/blob/master/Benchmarks\%20and\%20Evaluation.R}
}
\examples{
x <- AirPassengers
splitting_point <- round(2*length(x)/3)
x_train <- x[1:splitting_point]
h <- 5
x_hat <- thetaBenchmark(x_train,h)
}
