% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{M4_benchmarks}
\alias{M4_benchmarks}
\title{M4_benchmarks}
\usage{
M4_benchmarks(input, fh)
}
\arguments{
\item{input}{\itemize{
\item Input time series (numeric vector)
}}

\item{fh}{\itemize{
\item Forecasting horizon (numeric scalar)
}}
}
\value{
List containing:
\itemize{
\item{\code{Naive}: }{h-step forecast for the naive method (numeric vector - length h)}
\item{\code{NaiveSeasonal}: }{h-step forecast for the seasonal naive method (numeric vector - length h)}
\item{\code{SimpleES}: }{h-step forecast for the simple exponential smoothing method (numeric vector - length h)}
\item{\code{HoltWinters}: }{h-step forecast for the Holt-Winters method (numeric vector - length h)}
\item{\code{HoltWintersDamped}: }{h-step forecast for Holt-Winters damped method (numeric vector - length h)}
\item{\code{Theta}: }{h-step forecast for the theta method (numeric vector - length h)}
\item{\code{Combined}: }{h-step forecast for the combined method (numeric vector - length h)}
\item{\code{TimeNaive}: }{Computational time for the naive method (numeric scalar)}
\item{\code{TimeNaiveSeasonal}: }{Computational time for the seasonal naive method (numeric scalar)}
\item{\code{TimeSimpleES}: }{Computational time for the simple exponential smoothing method (numeric scalar)}
\item{\code{TimeHoltWinters}: }{Computational time for the Holt-Winters method (numeric scalar)}
\item{\code{TimeHoltWintersDamped}: }{Computational time for Holt-Winters damped method (numeric scalar)}
\item{\code{TimeTheta}: }{Computational time for the theta method (numeric scalar)}
\item{\code{TimeCombined}: }{Computational time for the combined method (numeric scalar)}
}
}
\description{
Function producing the univariate forecasts employing all the benchmark methods of the M4 competition.
Adapted from \url{https://github.com/M4Competition/M4-methods/blob/master/Benchmarks\%20and\%20Evaluation.R}
}
\examples{
x <- AirPassengers
splitting_point <- round(2*length(x)/3)
x_train <- x[1:splitting_point]
h <- 5
univariate_results <- M4_benchmarks(x_train,h)
}
