% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{multiplicativeSeasonalityDecomposition}
\alias{multiplicativeSeasonalityDecomposition}
\title{Auxiliary function for computing the multiplicative seasonality decomposition
From https://github.com/M4Competition/M4-methods/blob/master/Benchmarks\%20and\%20Evaluation.R}
\usage{
multiplicativeSeasonalityDecomposition(input, fh)
}
\arguments{
\item{input}{\itemize{
\item Input time series
}}

\item{fh}{\itemize{
\item Forecasting horizon
}}
}
\value{
des_input <- Deseasonalized input time series
SIout <- Additional deseasonalization parameters
}
\description{
Auxiliary function for computing the multiplicative seasonality decomposition
From https://github.com/M4Competition/M4-methods/blob/master/Benchmarks\%20and\%20Evaluation.R
}
