% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{multivariate_M4benchmarks}
\alias{multivariate_M4benchmarks}
\title{Quick and dirty multivariate implementation of the M4 benchmarks
https://github.com/M4Competition/M4-methods/blob/master/Benchmarks\%20and\%20Evaluation.R}
\usage{
multivariate_M4benchmarks(Xtr, H)
}
\arguments{
\item{Xtr}{\itemize{
\item Training data
}}

\item{H}{\itemize{
\item Forecasting horizon
}}
}
\value{
List of forecasts and excecution time for the different methods (Naive, Naive Seasonal, Naive 2, Simple Exponential smoothing ,Holt Winters, Holt Winters Damped, Theta, Comb)
Forecast are in a HxN(number of variables) matrix format, whereas the time is in a numeric format
}
\description{
Quick and dirty multivariate implementation of the M4 benchmarks
https://github.com/M4Competition/M4-methods/blob/master/Benchmarks\%20and\%20Evaluation.R
}
