% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{SeasonalityTest}
\alias{SeasonalityTest}
\title{Auxiliary function for computing a statistical test for seasonality
From https://github.com/M4Competition/M4-methods/blob/master/Benchmarks\%20and\%20Evaluation.R}
\usage{
SeasonalityTest(input, ppy)
}
\arguments{
\item{input}{\itemize{
\item Input time series
}}

\item{ppy}{\itemize{
\item Seasonality period
}}
}
\value{
test_seasonal <- Boolean value describing whether the time series is seasonal or not
}
\description{
Auxiliary function for computing a statistical test for seasonality
From https://github.com/M4Competition/M4-methods/blob/master/Benchmarks\%20and\%20Evaluation.R
}
