% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{Theta.classic}
\alias{Theta.classic}
\title{Auxiliary function for the Theta classic forecasting method.
From https://github.com/M4Competition/M4-methods/blob/master/Benchmarks\%20and\%20Evaluation.R}
\usage{
Theta.classic(input, fh)
}
\arguments{
\item{input}{\itemize{
\item Input time series
}}

\item{fh}{\itemize{
\item Forecasting horizon
}}
}
\value{
h-step forecast for the theta classic method
}
\description{
Auxiliary function for the Theta classic forecasting method.
From https://github.com/M4Competition/M4-methods/blob/master/Benchmarks\%20and\%20Evaluation.R
}
