% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{Theta.classic}
\alias{Theta.classic}
\title{Theta.classic}
\usage{
Theta.classic(input, fh)
}
\arguments{
\item{input}{\itemize{
\item Input time series (numeric vector)
}}

\item{fh}{\itemize{
\item Forecasting horizon (numeric scalar)
}}
}
\description{
Auxiliary function for the Theta classic forecasting method.
From \url{https://github.com/M4Competition/M4-methods/blob/master/Benchmarks\%20and\%20Evaluation.R}
}
