% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{thetaBenchmark}
\alias{thetaBenchmark}
\title{Auxiliary function for the theta forecasting method.
From https://github.com/M4Competition/M4-methods/blob/master/Benchmarks\%20and\%20Evaluation.R}
\usage{
thetaBenchmark(input, h)
}
\arguments{
\item{input}{\itemize{
\item Input time series
}}

\item{h}{\itemize{
\item Forecasting horizon
}}
}
\value{
h-step forecast for the theta classic method
}
\description{
Auxiliary function for the theta forecasting method.
From https://github.com/M4Competition/M4-methods/blob/master/Benchmarks\%20and\%20Evaluation.R
}
